+++ ionoscloud/data_source_dbaas_pgsql_cluster.go
@@ -9,7 +9,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	dbaas "github.com/ionos-cloud/sdk-go-dbaas-postgres"
 	dbaasService "github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/dbaas"
-	"log"
 )
 
 func dataSourceDbaasPgSqlCluster() *schema.Resource {
@@ -177,19 +176,31 @@ func dataSourceDbaasPgSqlReadCluster(ctx context.Context, d *schema.ResourceData
 			return diags
 		}
 	} else {
-		clusters, _, err := client.ListClusters(ctx, name.(string))
+		clusters, _, err := client.ListClusters(ctx, "")
 
 		if err != nil {
 			diags := diag.FromErr(fmt.Errorf("an error occurred while fetching dbaas clusters: %s", err.Error()))
 			return diags
 		}
 
+		var results []dbaas.ClusterResponse
+
 		if clusters.Items != nil && len(*clusters.Items) > 0 {
-			cluster = (*clusters.Items)[len(*clusters.Items)-1]
-			log.Printf("[WARN] %v clusters found matching the search criteria. Getting the latest datacenter from the list %v", len(*clusters.Items), *cluster.Id)
-		} else {
-			return diag.FromErr(errors.New("dbaas cluster not found"))
+			for _, clusterItem := range *clusters.Items {
+				if clusterItem.Properties != nil && clusterItem.Properties.DisplayName != nil && *clusterItem.Properties.DisplayName == name.(string) {
+					results = append(results, clusterItem)
+				}
+			}
 		}
+
+		if results == nil || len(results) == 0 {
+			return diag.FromErr(fmt.Errorf("no DBaaS cluster found with the specified name = %s", name))
+		} else if len(results) > 1 {
+			return diag.FromErr(fmt.Errorf("more than one DBaaS cluster found with the specified criteria name = %s", name))
+		} else {
+			cluster = results[0]
+		}
+
 	}
 
 	if err := dbaasService.SetDbaasPgSqlClusterData(d, cluster); err != nil {