+++ ionoscloud/data_source_server.go
@@ -8,7 +8,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	ionoscloud "github.com/ionos-cloud/sdk-go/v6"
-	"log"
 )
 
 func dataSourceServer() *schema.Resource {
@@ -661,18 +660,36 @@ func dataSourceServerRead(ctx context.Context, d *schema.ResourceData, meta inte
 		}
 	} else {
 		/* search by name */
-		servers, apiResponse, err := client.ServersApi.DatacentersServersGet(ctx, datacenterId.(string)).Depth(5).Filter("name", name.(string)).Execute()
+		servers, apiResponse, err := client.ServersApi.DatacentersServersGet(ctx, datacenterId.(string)).Depth(5).Execute()
 		logApiRequestTime(apiResponse)
 		if err != nil {
 			return diag.FromErr(fmt.Errorf("an error occurred while fetching servers: %w", err))
 		}
 
-		if servers.Items != nil && len(*servers.Items) > 0 {
-			server = (*servers.Items)[len(*servers.Items)-1]
-			log.Printf("[INFO] %v servers found matching the search criteria. Getting the latest server from the list %v", len(*servers.Items), *server.Id)
-		} else {
-			return diag.FromErr(fmt.Errorf("no server found with the specified name %s", name))
+		var results []ionoscloud.Server
+
+		if servers.Items != nil {
+			for _, s := range *servers.Items {
+				if s.Properties != nil && s.Properties.Name != nil && *s.Properties.Name == name.(string) {
+					/* server found */
+					server, apiResponse, err = client.ServersApi.DatacentersServersFindById(ctx, datacenterId.(string), *s.Id).Depth(4).Execute()
+					logApiRequestTime(apiResponse)
+					if err != nil {
+						return diag.FromErr(fmt.Errorf("an error occurred while fetching the server with ID %s: %w", *s.Id, err))
+					}
+					results = append(results, server)
+				}
+			}
 		}
+
+		if results == nil || len(results) == 0 {
+			return diag.FromErr(fmt.Errorf("no server found with the specified criteria: name = %s", name.(string)))
+		} else if len(results) > 1 {
+			return diag.FromErr(fmt.Errorf("more than one server found with the specified criteria: name = %s", name.(string)))
+		} else {
+			server = results[0]
+		}
+
 	}
 
 	var token = ionoscloud.Token{}