+++ ionoscloud/resource_networkloadbalancer_test.go
@@ -72,8 +72,8 @@ func TestAccNetworkLoadBalancerBasic(t *testing.T) {
 				),
 			},
 			{
-				Config:      testAccDataSourceNetworkLoadBalancerWrongName,
-				ExpectError: regexp.MustCompile(`no network load balancer found with the specified name`),
+				Config:      testAccDataSourceNetworkLoadBalancerWrongNameError,
+				ExpectError: regexp.MustCompile(`no network load balancer found with the specified criteria: name`),
 			},
 			{
 				Config: testAccCheckNetworkLoadBalancerConfigUpdate,
@@ -269,7 +269,7 @@ data ` + NetworkLoadBalancerResource + ` ` + NetworkLoadBalancerDataSourceByName
 }
 `
 
-const testAccDataSourceNetworkLoadBalancerWrongName = testAccCheckNetworkLoadBalancerConfigBasic + `
+const testAccDataSourceNetworkLoadBalancerWrongNameError = testAccCheckNetworkLoadBalancerConfigBasic + `
 data ` + NetworkLoadBalancerResource + ` ` + NetworkLoadBalancerDataSourceByName + ` {
   datacenter_id = ` + NetworkLoadBalancerResource + `.` + NetworkLoadBalancerTestResource + `.datacenter_id
   name			= "wrong_name"